requirements.txt
asgiref==3.8.1
Django==5.0.4
django-js-asset==3.1.2
django-mptt==0.17.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
sqlparse==0.5.0
tzdata==2024.1
